verify function to assign thetas, move thetas_to_thetas_dict to utils.py
